Alle Zugriffe vom JDO-Plugin auf das neue Storage-Interface geleitet.
git-svn-id: svn://localhost/twospot_svn/trunk@906 8d73be03-56cb-46cc-8948-98b08ae92ef2

+++ TwoSpotJdo/src/org/prot/storage/KeyCreator.java
@@ -11,7 +11,6 @@ import org.apache.hadoop.hbase.client.Put;
 import org.apache.hadoop.hbase.client.Result;
 import org.apache.hadoop.hbase.util.Bytes;
 import org.apache.log4j.Logger;
-import org.prot.stor.hbase.Key;
 import org.prot.storage.connection.ConnectionFactory;
 import org.prot.storage.connection.HBaseManagedConnection;
 import org.prot.storage.connection.StorageUtils;
@@ -38,7 +37,7 @@ public class KeyCreator
 		table.put(put);
 	}
 
-	public List<Key> fetchKey(String appId, int amount) throws IOException
+	public List<Key> fetchKey(String appId, long amount) throws IOException
 	{
 		HTable table = getSequenceTable();
